Dry up label pref comp
